refactor: Configuration 클래스 분리
